+++ backend/Api/Controllers/MonstersController.cs
@@ -11,28 +11,16 @@ namespace Api.Controllers
     public class MonstersController : ControllerBase
     {
         private readonly MonstersService _monstersService;
-        private readonly PatchService _patchService;
 
-        public MonstersController(MonstersService monstersService, PatchService patchService)
+        public MonstersController(MonstersService monstersService)
         {
             _monstersService = monstersService;
-            _patchService = patchService;
         }
 
         [HttpGet(nameof(Get))]
         public ActionResult<IEnumerable<Monster>> Get()
         {
             return Ok(_monstersService.Get());
-        }
-
-        [HttpPatch(nameof(Patch))]
-        public IActionResult Patch(CreationPatchRequest patch)
-        {
-            if (_patchService.PatchCreation(patch))
-            {
-                return Ok();
-            }
-            return BadRequest();
         }
     }
 }
