+++ frontend/src/components/CharacterSheet/NameField.js
@@ -1,7 +1,7 @@
 import { TextField, Tooltip } from "@mui/material";
 import { useEffect, useState } from "react"
 import { useDispatch, useSelector } from "react-redux";
-import { patchMonster } from "../../state/actions/monstersThunk";
+import { patchMonster } from "../../state/actions/creationThunk";
 import { creationState, monstersState } from "../../state/selectors";
 import field from '../../dictionaries/FieldNames.json';
