+++ frontend/src/components/CharacterSheet/NumberField.js
@@ -1,7 +1,7 @@
 import { TextField } from "@mui/material";
 import { useEffect, useState } from "react"
 import { useDispatch } from "react-redux";
-import { patchMonster } from "../../state/actions/monstersThunk";
+import { patchMonster } from "../../state/actions/creationThunk";
 import field from '../../dictionaries/FieldNames.json';
 import { patchCombatant } from "../../state/actions/combatThunk";
